Update to Products.Poi 2.2.5 and related versions.
We were using 2.0.2. Poi is compatible with Plone 4.3 since version 2.1.1
and has had a few important related changes since then.
With 2.0.2 you get an error when you add an issue on plone.org currently...
Please apply the upgrade steps in the Add-ons control panel.
This included updating the workflow settings, so this can take a while.